Moved raster and vector processes into separate files. Adjusted RequestParser accordingly to find the correct process class. Added a very simple celery task to run the RequestParser. Added celery configuration settings to gaia.dist.cf
