+++ gaia/conf/gaia.dist.cfg
@@ -3,3 +3,5 @@ fileio_paths: ../tests/data/
 output_path: ../tests/data/output
 tmp_dir: /tmp
 default_epsg: 3857
+celery_backend: amqp
+celery_broker: amqp://guest@localhost:5672//
