+++ gaia/parser.py
@@ -2,11 +2,19 @@ import argparse
 import traceback
 import json
 import os
+import inspect
 from six import string_types
 import gaia.formats
 import gaia.inputs
-import gaia.processes
+import gaia.processes_vector
+import gaia.processes_raster
 from gaia.core import GaiaException
+
+_process_r = [(x[0].replace('Process', ''), x[1]) for x in inspect.getmembers(
+    gaia.processes_raster, inspect.isclass) if x[0].endswith('Process')]
+_process_v = ([(x[0].replace('Process', ''), x[1]) for x in inspect.getmembers(
+    gaia.processes_vector, inspect.isclass) if x[0].endswith('Process')])
+_processes = dict(_process_r + _process_v)
 
 
 class GaiaRequestParser(object):
@@ -81,8 +89,7 @@ def create_process(name, parent=None):
     :return:
     """
     try:
-        class_name = '{}Process'.format(name[0].capitalize() + name[1:])
-        return getattr(gaia.processes, class_name)(parent=parent)
+        return _processes[name[0].capitalize() + name[1:]](parent=parent)
     except AttributeError:
         raise GaiaException(traceback.format_exc())
 
@@ -94,7 +101,7 @@ def parse_request(process, request_json):
     :param request_json: The process configuration in JSON format
     :return: A GaiaProcess object
     """
-    parser = GaiaRequestParser(process, data=request_json, parse=True)
+    parser = GaiaRequestParser(process, request_json, parse=True)
     parser.process.compute()
     return parser.process
