+++ gaia/processes_vector.py
@@ -1,6 +1,4 @@
 import logging
-import uuid
-import os
 import numpy as np
 import pandas as pd
 try:
@@ -8,50 +6,12 @@ try:
 except ImportError:
     from osgeo import osr
 from geopandas import GeoDataFrame
-from gaia.core import get_abspath, config, GaiaException
-from gaia.inputs import VectorFileIO, RasterFileIO, reproject
 import gaia.formats as formats
-from gaia.gdal_functions import gdal_clip, gdal_calc
-
+from gaia.core import GaiaException
+from gaia.inputs import VectorFileIO, reproject
+from gaia.processes_base import GaiaProcess
 
 logger = logging.getLogger('gaia.processes')
-
-
-class GaiaProcess(object):
-    """
-    Defines a process to run on geospatial inputs
-    """
-
-    # TODO: Enforce required inputs and args
-    required_inputs = tuple()
-    required_args = tuple()
-
-    def __init__(self, inputs=None, output=None, args=None, parent=None):
-        self.inputs = inputs
-        self.output = output
-        self.args = args
-        self.parent = parent
-        self.id = str(uuid.uuid4())
-
-    def compute(self):
-        previous_input = None
-        for input in self.inputs:
-            if input.data is None:
-                input.read()
-            if previous_input:
-                if input.get_epsg() != previous_input.get_epsg():
-                    reproject(input, previous_input.epsg)
-            previous_input = input
-
-    def purge(self):
-        self.output.delete()
-
-    def get_outpath(self, uri=config['gaia']['output_path']):
-        ids_path = '{}/{}'.format(
-            self.parent, self.id) if self.parent else self.id
-        return get_abspath(
-            os.path.join(uri, ids_path,
-                         '{}{}'.format(self.id, self.default_output[0])))
 
 
 class BufferProcess(GaiaProcess):
@@ -90,38 +50,6 @@ class BufferProcess(GaiaProcess):
         if original_projection:
             reproject(self.output, original_projection)
             self.output.write()
-
-
-class SubsetProcess(GaiaProcess):
-    """
-    Generates a raster dataset representing the portion of the input raster
-    dataset that is contained within a vector polygon.
-    """
-    required_inputs = (('clip', formats.JSON), ('raster', formats.RASTER))
-    default_output = formats.RASTER
-
-    def __init__(self, **kwargs):
-        """
-        Create a process to subset a raster by a vector polygon
-        :param clip_io: IO object containing vector polygon data
-        :param raster_io: IO object containing raster data
-        :param kwargs:
-        :return: SubsetProcess object
-        """
-        super(SubsetProcess, self).__init__(**kwargs)
-        if not self.output:
-            self.output = RasterFileIO(name='result',
-                                       uri=self.get_outpath())
-
-    def compute(self):
-        super(SubsetProcess, self).compute()
-        clip_df = self.inputs[0].read()
-        raster_img = self.inputs[1].read()
-        # Merge all features in vector input
-        raster_output = self.output.uri
-        self.output.create_output_dir(raster_output)
-        clip_json = clip_df.geometry.unary_union.__geo_interface__
-        self.output.data = gdal_clip(raster_img, raster_output, clip_json)
 
 
 class WithinProcess(GaiaProcess):
@@ -295,41 +223,3 @@ class DistanceProcess(GaiaProcess):
         self.output.data = first_df
         self.output.write()
         logger.debug(self.output)
-
-
-class RasterMathProcess(GaiaProcess):
-    """
-    Performs raster math/algebra based on supplied arguments.
-    Inputs should consist of at least one raster IO object.
-    Required arg is 'calc', an equation for the input rasters.
-    Example: "A + B / (C * 2.4)".  The letters in the equation
-    should correspond to the names of the inputs.
-    """
-    required_inputs = (('A', formats.RASTER),)
-    required_args = ('calc',)
-    optional_args = ('bands', 'nodata', 'allBands', 'output_type')
-    default_output = formats.RASTER
-
-    def __init__(self, **kwargs):
-        super(RasterMathProcess, self).__init__(**kwargs)
-        if not self.output:
-            self.output = RasterFileIO(name='result',
-                                       uri=self.get_outpath())
-
-    def compute(self):
-        super(RasterMathProcess, self).compute()
-
-        calculation = self.args['calc']
-        rasters = [x.read() for x in self.inputs]
-        bands = self.args.get('bands' or None)
-        nodata = self.args.get('nodata' or None)
-        all_bands = self.args.get('allBands' or None)
-        otype = self.args.get('output_type' or None)
-        self.output.create_output_dir(self.output.uri)
-        self.output.data = gdal_calc(calculation,
-                                     self.output.uri,
-                                     rasters,
-                                     bands=bands,
-                                     nodata=nodata,
-                                     allBands=all_bands,
-                                     output_type=otype)
