+++ requirements.txt
@@ -19,3 +19,4 @@ argparse==1.3.0
 geopandas==0.1.1
 pillow==3.1.1
 gdal==1.10
+celery==3.1.20
